+++ Changelog.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Allow overriding platform release version using env var in [#975](https://github.com/PyO3/maturin/pull/975)
 * Fix `maturin develop` for arm64 Python on M1 Mac when default toolchain is x86_64 in [#980](https://github.com/PyO3/maturin/pull/980)
 * Add `--repository` option to `maturin upload` command in [#987](https://github.com/PyO3/maturin/pull/987)
+* Only lookup bundled Python sysconfig when interpreters aren't specified as file path in [#988](https://github.com/PyO3/maturin/pull/988)
 
 ## [0.12.20] - 2022-06-15